extract time zone logic into separate file from time zone values
