+++ activesupport/lib/active_support/core_ext/date_time/conversions.rb
@@ -2,7 +2,7 @@ require 'date'
 require 'active_support/inflector/methods'
 require 'active_support/core_ext/time/conversions'
 require 'active_support/core_ext/date_time/calculations'
-require 'active_support/values/time_zone'
+require 'active_support/time_zone'
 
 class DateTime
   # Convert to a formatted string. See Time::DATE_FORMATS for predefined formats.
