+++ activesupport/lib/active_support/time.rb
@@ -1,7 +1,7 @@
 module ActiveSupport
   autoload :Duration, 'active_support/duration'
   autoload :TimeWithZone, 'active_support/time_with_zone'
-  autoload :TimeZone, 'active_support/values/time_zone'
+  autoload :TimeZone, 'active_support/time_zone'
 end
 
 require 'date'
